Extract test item pipeline module
